requirements.txt
Flask>=2.3.2
pyotp>=2.8.0
PyJWT>=2.6.0
passlib>=1.7.4
cryptography>=40.0.2
Unidecode>=1.3.6
MarkupSafe>=2.1.2
Werkzeug>=2.3.3
alembic>=1.10.4
SQLAlchemy>=2.0.12
Flask-Migrate>=4.0.4
Flask-RESTful>=0.3.9
Flask-Cors>=3.0.10
Flask-WTF>=1.1.1
Flask-Mail>=0.9.1
Flask_httpauth>=4.8.0
flask_limiter>=3.3.1
flask_sslify>=0.1.5
pwnedpasswords>=2.0.0
decouple>=0-7
defusedxml>=0.7.1
validators >= 0.20.0
requests >= 2.31.0
psycopg2 >= 2.9.6
gunicorn